+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Updated/Enhanced Rules and Bug Fixes
  - #688: **SpaceAfterMethodCallName** rule: Fix false positives. NOTE: No longer check method calls without parentheses.
  - #703: **UnnecessaryPublicModifier** rule: Fix PatternSyntaxException if method name contains special (regex) characters.
  - #702: **Indentation** rule: Fix false positive for try-with-resources. Also skip generated code and single-line blocks.
+ - #692: **UnusedPrivateField**: Ignore classes annotated with @Entity. Add *ignoreClassesAnnotatedWithNames* property.
 
 Build, Infrastructure and Tests
 - #709: Publish Groovy 4.x-specific jar for CodeNarc. Upgrade to GMetrics 2.1.0.